+++ src/detail/llvm_helpers.cpp
@@ -39,6 +39,7 @@
 
 #include <llvm/Analysis/VectorUtils.h>
 #include <llvm/Config/llvm-config.h>
+#include <llvm/IR/Argument.h>
 #include <llvm/IR/Attributes.h>
 #include <llvm/IR/BasicBlock.h>
 #include <llvm/IR/Constant.h>
@@ -58,6 +59,7 @@
 #include <llvm/IR/Value.h>
 #include <llvm/Support/Alignment.h>
 #include <llvm/Support/Casting.h>
+#include <llvm/Support/ModRef.h>
 #include <llvm/Support/raw_ostream.h>
 
 #if defined(HEYOKA_HAVE_REAL128)
@@ -168,10 +170,10 @@ const auto type_map = []() {
             return ptr;
         }
 
-        auto *ret = llvm::StructType::create(
-            {to_external_llvm_type<mpfr_prec_t>(c), to_external_llvm_type<mpfr_sign_t>(c),
-             to_external_llvm_type<mpfr_exp_t>(c), llvm::PointerType::getUnqual(to_external_llvm_type<mp_limb_t>(c))},
-            "heyoka.real");
+        auto *ret
+            = llvm::StructType::create({to_external_llvm_type<mpfr_prec_t>(c), to_external_llvm_type<mpfr_sign_t>(c),
+                                        to_external_llvm_type<mpfr_exp_t>(c), llvm::PointerType::getUnqual(c)},
+                                       "heyoka.real");
 
         assert(ret != nullptr);
         assert(llvm::StructType::getTypeByName(c, "heyoka.real") == ret);
@@ -1150,7 +1152,7 @@ llvm::Value *ext_load_vector_from_memory(llvm_state &s, llvm::Type *tp, llvm::Va
 
         // Load in a local variable the input pointer to the limbs.
         auto *limb_ptr_ptr = builder.CreateInBoundsGEP(real_t, ptr, {builder.getInt32(0), builder.getInt32(3)});
-        auto *limb_ptr = builder.CreateLoad(llvm::PointerType::getUnqual(limb_t), limb_ptr_ptr);
+        auto *limb_ptr = builder.CreateLoad(llvm::PointerType::getUnqual(context), limb_ptr_ptr);
 
         // Load and insert the limbs.
         for (std::size_t i = 0; i < nlimbs; ++i) {
@@ -1244,7 +1246,7 @@ void ext_store_vector_to_memory(llvm_state &s, llvm::Value *ptr, llvm::Value *ve
 
         // Load in a local variable the output pointer to the limbs.
         auto *out_limb_ptr_ptr = builder.CreateInBoundsGEP(real_t, ptr, {builder.getInt32(0), builder.getInt32(3)});
-        auto *out_limb_ptr = builder.CreateLoad(llvm::PointerType::getUnqual(limb_t), out_limb_ptr_ptr);
+        auto *out_limb_ptr = builder.CreateLoad(llvm::PointerType::getUnqual(context), out_limb_ptr_ptr);
 
         // Store the limbs.
         for (std::size_t i = 0; i < nlimbs; ++i) {
@@ -2623,13 +2625,13 @@ llvm::Function *llvm_add_csc(llvm_state &s, llvm::Type *scal_t, std::uint32_t n,
     const auto fname = fmt::format("heyoka_csc_degree_{}_{}", n, llvm_mangle_type(tp));
 
     // The function arguments:
+    //
     // - pointer to the return value,
     // - pointer to the array of coefficients.
-    // NOTE: both pointers are to the scalar counterparts
-    // of the vector types, so that we can call this from regular
-    // C++ code. The second pointer is to an external type.
-    const std::vector<llvm::Type *> fargs{llvm::PointerType::getUnqual(builder.getInt32Ty()),
-                                          llvm::PointerType::getUnqual(ext_fp_t)};
+    //
+    // NOTE: both pointers are to the scalar counterparts of the vector types, so that we can call this from regular C++
+    // code. The second pointer is to an external type.
+    const std::vector<llvm::Type *> fargs(2, llvm::PointerType::getUnqual(context));
 
     // Try to see if we already created the function.
     auto *f = md.getFunction(fname);
@@ -2649,13 +2651,13 @@ llvm::Function *llvm_add_csc(llvm_state &s, llvm::Type *scal_t, std::uint32_t n,
         // Fetch the necessary function arguments.
         auto *out_ptr = f->args().begin();
         out_ptr->setName("out_ptr");
-        out_ptr->addAttr(llvm::Attribute::NoCapture);
+        llvm_add_no_capture_argattr(s, out_ptr);
         out_ptr->addAttr(llvm::Attribute::NoAlias);
         out_ptr->addAttr(llvm::Attribute::WriteOnly);
 
         auto *cf_ptr = f->args().begin() + 1;
         cf_ptr->setName("cf_ptr");
-        cf_ptr->addAttr(llvm::Attribute::NoCapture);
+        llvm_add_no_capture_argattr(s, cf_ptr);
         cf_ptr->addAttr(llvm::Attribute::NoAlias);
         cf_ptr->addAttr(llvm::Attribute::ReadOnly);
 
@@ -4002,6 +4004,44 @@ HEYOKA_DLL_PUBLIC void llvm_assert([[maybe_unused]] llvm_state &s, [[maybe_unuse
                                   bld.getInt64(boost::numeric_cast<std::uint64_t>(loc.column())), file_name,
                                   function_name});
         });
+
+#endif
+}
+
+// Helper to add the nocapture attribute to a pointer function argument. The syntax changes in LLVM 21, hence the need
+// for a wrapper.
+void llvm_add_no_capture_argattr([[maybe_unused]] llvm_state &s, llvm::Argument *arg)
+{
+    assert(arg != nullptr);
+    assert(arg->getType()->isPointerTy());
+
+#if LLVM_VERSION_MAJOR <= 20
+
+    arg->addAttr(llvm::Attribute::NoCapture);
+
+#else
+
+    arg->addAttr(llvm::Attribute::getWithCaptureInfo(s.context(), llvm::CaptureInfo::none()));
+
+#endif
+}
+
+// Helper to check if the input type is an IEEE-like floating-point type.
+//
+// NOTE: LLVM<=20 had an isIEEE() method for this, but it got slightly changed in LLVM 21 so that now it is called
+// isIEEELikeFPTy() and it *excludes* 80-bit extended precision. For our internal use, we want to consider 80-bit
+// extended precision as IEEE-like.
+bool llvm_is_ieee_like_fp(llvm::Type *tp)
+{
+    assert(tp != nullptr);
+
+#if LLVM_VERSION_MAJOR <= 20
+
+    return tp->isFloatingPointTy() && tp->isIEEE();
+
+#else
+
+    return tp->isFloatingPointTy() && (tp->isIEEELikeFPTy() || tp->isX86_FP80Ty());
 
 #endif
 }